feat(neovim): update plugins and add smart-open
This commit updates the versions of several Neovim plugins in the
lazy-lock.json file. It also adds the smart-open plugin to the
configuration, which provides a smarter way to open files in Neovim.

In addition, the commit introduces new commands for navigating to
commands, files, and symbols, and for performing live grep. It also
disables automatic plugin updates.

Finally, the commit makes some changes to the Wezterm configuration,
including enabling some commands that were previously commented out.